fix (service, routes): create separate service class for uploading image on aws, add route for user logs

+++ app/Http/Controllers/Admin/BookController.php
@@ -7,15 +7,15 @@ use App\Http\Requests\StoreBookRequest;
 use App\Http\Requests\UpdateBookRequest;
 use App\Http\Resources\BookResource;
 use App\Models\Book;
-use App\Models\FavoriteBook;
+use App\Services\UploadImageService;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Log;
-use Illuminate\Support\Facades\Storage;
 use Inertia\Inertia;
 
 class BookController extends Controller
 {
+    public function __construct(public UploadImageService $uploadImageService)
+    {
+    }
     /**
      * Display a listing of the resource.
      */
@@ -47,9 +47,8 @@ class BookController extends Controller
     {
         $data = $request->validated();
 
-        $uploadPath = Storage::disk('s3')->putFile('books', $data['poster']);
-        $data['poster_short_url'] = $uploadPath;
-        $data['poster'] = Storage::disk('s3')->url($uploadPath);
+        $imageData = $this->uploadImageService->uploadImage($data['poster'], 'books');
+        $data = array_merge($data, $imageData);
 
         $book = Book::create($data);
 
@@ -81,17 +80,16 @@ class BookController extends Controller
     {
         $data = $request->validated();
         
-        if($data['poster'] !== $book->poster && null !== $book->poster_short_url) {
-            Storage::disk('s3')->delete($book->poster_short_url);
-        }
-        if($request->file('poster')) {
-            $uploadPath = Storage::disk('s3')->putFile('books', $data['poster']);
-            $data['poster_short_url'] = $uploadPath;
-            $data['poster'] = Storage::disk('s3')->url($uploadPath);
+        if ($request->hasFile('poster')) {
+            if ($book->poster_short_url) {
+                $this->uploadImageService->deleteImage($book->poster_short_url);
+            }
+            
+            $imageData = $this->uploadImageService->uploadImage($data['poster'], 'books');
+            $data = array_merge($data, $imageData);
         }
 
         $book->update($data);
-        $book->refresh();
         
         return redirect()
             ->route('admin.books.edit', ['book' => $book])
@@ -103,7 +101,10 @@ class BookController extends Controller
      */
     public function destroy(Book $book)
     {
-        Storage::disk('s3')->delete($book->poster_short_url);
+        $isDeleted = $this->uploadImageService->deleteImage($book->poster_short_url);
+        if(false === $isDeleted) {
+            return back()->withErrors(['image' => 'Image wasnt deleted']);
+        }
         $book->delete();
 
         return back()->with('success', 'You remove book from your library!');
